+++ NEWS.md
@@ -1,5 +1,7 @@
 # tidycovid19 0.0.0.9000
 
+* `*_covid19_spread()`: Added option for averaged daily changes
+
 * `*_covid19_spread()`: Added option for linear Y-axis
 
 * `shiny_covid19_spread()`: Added option to copy plot code to clipboard